requirements.txt
torch
requests
python-telegram-bot
python-dotenv
transformers
numpy